// src/contracts/generated/csi11/EmailLinkPreviewResultV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EmailChallengeKindV1 } from "./EmailChallengeKindV1";

export type EmailLinkPreviewResultV1 = { "kind": "ready", schemaVersion: number, challengeKind: EmailChallengeKindV1, email: string, productNameMessageKey: string, preview: string, expiresAt: string, };